Changed sequence packaged structure and reduced duplicate codes for array. (#19)
