+++ data_structures/sequence/linked_list/custom_linked_list.py
