+++ data_structures/sequence/linked_list/doubly_linked_list.py
