+++ data_structures/sequence/queue_stack/linked_queue.py
@@ -7,7 +7,7 @@ practical usage.
 """
 from typing import TypeVar, Optional, Sequence
 from .custom_queue import CustomQueue
-from .custom_linked_list import LinkedListMixin
+from .. import LinkedListMixin
 
 
 GT = TypeVar('GT')
