+++ data_structures/sequence/queue_stack/queued_stack.py
